Create index.html
Index Page (Home Page):
This page is the main introduction to the “Being Black” website. It includes the title of the site, a short welcome message, and a banner image. The home page explains that the website is about Black beauty, culture, and tips. It also has simple navigation links that let users move to the About, Blog, and Contact pages.