Create blog.html
In this step, I created my blog page. I added a section for future posts and kept the topic focused on Black beauty. I also put the newspaper-style box around the content so the page matches the rest of my website. This gives the blog a clean and simple layout while keeping the nude theme.